Update amounts in app.js
Some testnet faucets only issue a fraction of an ether. Reducing the price ensures that the transaction will still push through just in case it was done on the test network (although this is not required in the project). The "gwei" unit was used for backwards compatibility just in case an older version of web3 was used.